Extract backtest Running from BacktestView
